Fix: types in eePromTools

+++ lib/eePromTools/eePromTools.cpp
@@ -1,10 +1,10 @@
 #include "eepromTools.h"
 
-String eepromRead(u_int8_t startAddress, u_int8_t endAddress) {
+String eepromRead(unsigned int startAddress, unsigned int endAddress) {
     EEPROM.begin(512);
     String value = "";
     char current;
-    for(int i = startAddress; i < endAddress; ++i){
+    for(unsigned int i = startAddress; i < endAddress; ++i){
         current = EEPROM.read(i);
         if(current == '\0') break;
         value += current;
@@ -14,9 +14,9 @@ String eepromRead(u_int8_t startAddress, u_int8_t endAddress) {
     return value;
 }
 
-String eepromWrite(u_int8_t startAddress, String value) {
+String eepromWrite(unsigned int startAddress, String value) {
     EEPROM.begin(512);
-    for(int i = 0; i < value.length(); ++i){
+    for(unsigned int i = 0; i < value.length(); ++i){
         EEPROM.write(startAddress + i, value[i]);
     }
     EEPROM.write(startAddress + value.length(), '\0');
